third_party/blink/web_tests/svg/dynamic-updates/SVGFESpecularLightingElement-svgdom-specularConstant-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'specularConstant' property of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.specularConstant.baseVal is 4
PASS specularElement.specularConstant.baseVal is 1
PASS successfullyParsed is true

TEST COMPLETE
